Readbits binarises only the sample points
The speed gain is just micro-optimisation, but the detection is increased
a lot, because OTSU finds a threshold to classify only the sampled pixels
rather than the whole tag image.

Readbits was also improved by computing the sampling positions directly from
the corners of the tag rather than an estimation of the corners of the data
matrix.