More feedback on which tag fails the tests
Extracted the metadata about the test images (i.e. filename and which tags are
present on the image) into a separate file.
Added the list of tags that are known to be undetected, in order to easily get
a feedback on which ones are new false negatives, and which ones are new true
positives.