Kept only one constructor for Objects
The constructor only requires a default marker size. The filename of the
object configuration file is empty by default, meaning that the pose of all the
tags will be estimated independently. If the size is negative, only tags
explicitely marked as "keep" in the config file will be estimated.

The constructor also give default values to the camera matrix and the
distortion coefficients (respectively: an identity matrix, and a null vector).
The calibration parameter can be set with a subsequent call to setCalibration.