Simplified the API
Kepts only one public header, containing 2 classes:
Chilitags performs the detection on images,
and Chilitags3D uses it to estimate the 3d pose.
They are not functors anymore.
Other utilities such as DrawTag or Codec are accessible via Chilitags.
The persistence filter is embedded in Chilitags, controllable via a parameter.
A boolean controls which tags are 3d-estimated instead of testing whether the
default marker corners are empty.
The boolean is an optional parameter of the readConfiguration method,
because we don't need it before (without a configuration file, we want to
estimate the position of all the tags)
The camera calibration parsing is now part of into Chilitags3D