Do not specify anymore required OpenCV components
Commit 50a8a7b5ba tried to be intelligent and allow compilation without
OpenCV's highgui and ts when they were not needed.

But this broke the compilation with OpenCV 2.4 because the
find_package(OpenCV QUIET COMPONENTS highgui ts) call would overwrite
the previous OpenCV configuration (which is an OpenCV bug, as far as I
understand).

To simplify the situation, we now only call
find_package(OpenCV REQUIRED): all available OpenCV headers/libraries will be
included (which should only marginally impact compilation time), and
guards check that highgui/ts are available if needed.